requirements.txt
alembic==1.13.1
antiorm==1.2.1
apispec==6.3.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
db==0.1.1
db-sqlite3==0.0.1
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pluggy==1.3.0
PyJWT==2.8.0
pytest==7.4.4
python-chess==1.999
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
webargs==8.3.0
Werkzeug==3.0.1
zstandard==0.22.0
psycopg2-binary
psycopg2

